Add separation of decision maker
