+++ packages/hashes.csv
@@ -97,7 +97,7 @@ fetchai/skills/simple_service_registration,QmVuTkmxFaddeTznwmDY9C345bcUGiTHKJxZ7
 fetchai/skills/simple_service_search,QmdjAShNy1RcH9X7hYRJ1BpkYBi5M6zc9XyukXjF8Eyrfa
 fetchai/skills/tac_control,QmRBK7oB4WPSBr2GVVMWdUN9dB8pk3q7D8uDbroGgJ1fHS
 fetchai/skills/tac_control_contract,QmVaFcEb6puB6aEfQUz6oAiEQ3eZf7JJWVQLyQtLX28um5
-fetchai/skills/tac_negotiation,QmUUeV6qzVMrpvZrPbCzqHRCza9MvUBjvsJHnXCEseALVy
+fetchai/skills/tac_negotiation,QmSiQ6cf6knqfmzAHtMGYru1EMn5Rffn4pKVuePNnfmutf
 fetchai/skills/tac_participation,QmQjv8SRjis1NFpGC2gXYUoJ8tNWpDrFqQ9QpLJPC2hEhh
 fetchai/skills/thermometer,QmaLJyZQQefcvCdUC8N6uP8TW21J6HCG1irU3nqmdzC7dW
 fetchai/skills/thermometer_client,QmZgtnVrcBtvoiEXnbmZE1yFsheTn9dB4DS6JV7GoL5Uu9
