+++ tests/test_decision_maker/test_ownership_state.py
@@ -21,7 +21,7 @@
 
 import pytest
 
-from aea.decision_maker.default import OwnershipState
+from aea.decision_maker.gop import OwnershipState
 from aea.helpers.transaction.base import Terms
 
 from tests.conftest import ETHEREUM
